💄 Split helper functions for game mechanics and refactor rule management
